Omaha #3410 changed lon/lat fields to floats in sfc and air obs location

Former-commit-id: aa991c922354f0a67fac08fee456ca60a30de968 [formerly cdb0a72fcb2345564f2b944fdecd7869fd498767]
Former-commit-id: 74a6c3dd20898e9bd4484376ea4e4e471bdec9af

+++ edexOsgi/com.raytheon.edex.plugin.profiler/src/com/raytheon/edex/plugin/profiler/decoder/ProfilerDataAdapter.java
@@ -62,6 +62,7 @@ import com.raytheon.uf.edex.decodertools.core.IDecoderConstants;
  * Dec 03, 2013  2537     bsteffen    Switch logger to ufstatus.
  * May 14, 2014  2536     bclement    moved WMO Header to common, removed TimeTools usage
  *                                     added breaks/default to switch
+ * Jul 23, 2014 3410      bclement    location changed to floats
  * 
  * </pre>
  * 
@@ -212,7 +213,7 @@ public class ProfilerDataAdapter {
                 ProfilerSite site = profilers.get(wmoBSN);
                 if (site != null) {
                     location.setStationId(site.getStationId());
-                    location.assignLocation(lat, lon);
+                    location.assignLocation(lat.floatValue(), lon.floatValue());
                     location.setLocationDefined(false);
                     if (stationHeight != IDecoderConstants.VAL_MISSING) {
                         location.setElevation(stationHeight);
